[Cleanup] Split ScratchpadBindingToken and DFBBindingToken into their own headers
The Metal 2.0 binding tokens are what the auto-generated
kernel_bindings_generated.h needs to name a kernel's bindings. Two of the
four token types were defined inside the full implementation header for
their resource, so the generated header had to pull in the entire
implementation just to declare a token.

SemaphoreBindingToken (api/dataflow/semaphore_binding_token.h) and
TensorBindingToken (api/tensor/tensor_binding_token.h) already live in
their own headers. This makes the other two match:

  - DFBBindingToken        -> api/dataflow/dfb_binding_token.h
  - ScratchpadBindingToken -> api/scratchpad_binding_token.h

Both new headers depend on <cstdint> and nothing else; the scratchpad one
forward-declares Scratchpad<T> for its friend declaration rather than
including it. Verified self-contained, and verified that a generated-header
shape built from the three light token headers alone compiles.

No behaviour change: dataflow_buffer.h and scratchpad.h now include the
token header they used to define, so every existing include site is
unaffected.
